requirements.txt
Django==1.10.4
Pillow
MySQL-python
django-crispy-forms
python-dateutil